### 2023-03-17 #91 @sampottinger
afscgap: Ocean health tools for NOAA AFSC GAP species presence data [presubmission]
Submitting Author: Sam Pottinger (@sampottinger)
Package Name: `afscgap`
One-Line Description of Package: Community contributed Python-based tools for working with public bottom trawl surveys data from the NOAA Alaska Fisheries Science Center Groundfish Assessment Program ([NOAA AFSC GAP](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program)).
Repository Link (if existing): https://github.com/SchmidtDSE/afscgap

---

Python-based tool set for interacting with [bottom trawl surveys](https://www.fisheries.noaa.gov/alaska/commercial-fishing/alaska-groundfish-bottom-trawl-survey-data) from the [Ground Fish Assessment Program (GAP)](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program). This provides information about where certain species were seen and when under what conditions, information useful for research in ocean health.

It offers:

 - Pythonic access to the official [NOAA AFSC GAP API service](https://www.fisheries.noaa.gov/foss/f?p=215%3A28).
 - Memory-efficient tools for inference of the ["negative" observations](https://github.com/SchmidtDSE/afscgap#absence-vs-presence-data) not provided by the API service but required for some common types of analysis.
 - [Visualization tools](https://app.pyafscgap.org) for quickly exploring and creating comparisons within the dataset that provide an "on-ramp" to deeper analysis. The visual analytics components aim to serve both expert programmers and audiences with limited programming experience.

Note that GAP is an excellent dataset produced by the [Resource Assessment and Conservation Engineering (RACE) Division](https://www.fisheries.noaa.gov/about/resource-assessment-and-conservation-engineering-division) of the [Alaska Fisheries Science Center (AFSC)](https://www.fisheries.noaa.gov/about/alaska-fisheries-science-center) as part of the National Oceanic and Atmospheric Administration's Fisheries organization ([NOAA Fisheries](https://www.fisheries.noaa.gov/)).

Additional information at [https://pyafscgap.org/](https://pyafscgap.org/).

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

- Domain Specific & Community Partnerships

This library supports retrieval of data from the official NOAA AFSC GAP REST API service but, as that service on its own is often not sufficient due to certain data limitations, it also offers negative inference as required for a number of common types of investigations (hence data processing). Finally, given the needs of the community and the vast breadth of the dataset, it offers a community application to explore the data which, in turn, can generate Python code to help users get started with continued analysis within their own scripts.

This project largely benefits scientific researchers in the ocean health space as this dataset is useful for fisheries management, biodiversity research, and marine science more generally. An example of what this analysis may look like is provided in our [example notebook](https://mybinder.org/v2/gh/SchmidtDSE/afscgap/main?urlpath=/tree/index.ipynb) hosted on mybinder.org.

We are not aware of other Python packages working with the AFSC GAP dataset.

We've found the visualization tool to be an important part of this broader toolset, useful for lowering barrier to entry, starting investigations, and inviting a broader group of folks into discourse. That said, while related and subject to automated tests, CI / CD, and doc requirements, pyopensci may choose to consider `afscgapviz` as a form of documentation from the perspective of the python library.

[Conduct]: https://www.pyopensci.org/governance/CODE_OF_CONDUCT
