+++ README.md
@@ -1,5 +1,5 @@
 # AFSC GAP for Python
-Python-based tool chain for working with the public bottom trawl surveys data from the [NOAA AFSC GAP](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program).
+Python-based tool chain for working with the public bottom trawl surveys data from the [NOAA AFSC GAP](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program). This provides information about where certain species were seen and when under what conditions, information useful for reseach in ocean health.
 
 | Group | Badges |
 |-------|--------|
