+++ afscgapviz/load_deps.sh
@@ -1,0 +1,5 @@
+mkdir static/js_third_party
+wget https://d3js.org/d3.v7.min.js -O static/js_third_party/d3.min.js
+wget https://cdnjs.cloudflare.com/ajax/libs/PapaParse/5.4.0/papaparse.min.js -O static/js_third_party/papaparse.min.js
+wget https://cdnjs.cloudflare.com/ajax/libs/qunit/2.19.4/qunit.min.css -O static/test/qunit.css
+wget https://cdnjs.cloudflare.com/ajax/libs/qunit/2.19.4/qunit.min.js -O static/test/qunit.min.js
