### 2023-03-20 d8aa16c
Formalize error in build database script.
