### 2023-03-21 #91 @NickleDave
👋 Welcome @sampottinger!

Thank you for filling out a presubmission enquiry.

afscgap is definitely in scope. Please do go ahead and make a full submission.

You've done a great job of filling out the form here (I actually thought this was a full submission already, hence the delay while I started doing checks, sorry about that 😳 ). If there are other maintainers please be sure to include them and please have everyone fill out the pre-submission survey.

Once you do so I'll close this issue, start editor checks, and proceed with finding an editor.
Thanks! afscgap looks like a really useful tool, happy to get you a review
