### 2023-03-22 #91 @sampottinger
Hey @NickleDave! We completed the survey and I have a draft of the full submission ready. However, I think I misunderstood the docs regarding JOSS submissions through pyopensci. I wanted to confirm if the paper.md should be on main at time of submission? Thanks!
