+++ paper.md
@@ -1,0 +1,83 @@
+---
+title: 'Pyafscgap.org: Python tools for NOAA AFSC GAP'
+tags:
+  - Python
+  - fishery
+  - alaska
+  - groundfish
+  - biodiversity
+  - food
+  - visualization
+authors:
+  - name: A Samuel Pottinger
+    orcid: 0000-0002-0458-4985
+    affiliation: 1
+  - name: Giulia Zarpellon
+    orcid: 0000-0002-9122-4709
+    affiliation: 1
+affiliations:
+ - name: University of California Berkeley
+   index: 1
+date: 21 March 2023
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center (NOAA AFSC) runs the [Groundfish Assessment Program](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program) which produces marine longitudinal catch data [@afscgap]. These "hauls" report where species are found and in what quantities, empowering biodiversity research and fisheries management [@pacificcod]. Increasing accessibility of these important data through a suite of tools for individuals of diverse programming experience, Pyafscgap.org offers not just easier access to the REST API through ORDS compilation but provides both memory-efficient algorithms for "zero-catch inference" and interactive visual analytics tools [@ords]. Altogether, this toolset supports investigatory tasks not easily executable using the API service alone and, leveraging game and information design, offers these data to a broader audience.
+
+# Statement of need
+
+Pyafscgap.org reduces barriers for use of AFSC GAP data, offering open source solutions for addressing the dataset's presence-only nature, use of proprietary databases, size, and complexity [@inport].
+
+## Developer needs
+The `afscgap` package offers easier developer access to the official REST service with automated pagination, query language compilation, and formalized documented types. Together, these tools enable Python developers to use familiar patterns to interact with these data like type checking, standard documentation, and other Python data-related libraries while freeing programs from structures specific to closed-source Oracle REST Data Services (ORDS) [@ords].
+
+That being said, access to the API alone cannot support some investigations as the API provides "presence-only" data [@inport]. Many types of analysis like aggregating species catch per unit effort by geohash require information not just about where a species was present but also where it was not [@geohash]. In other words, while the presence-only dataset may provide a total weight or specimen count, the total area swept for a region may not necessarily be easily available but required [@notebook]. The `afscgap` Python package can, with memory efficiency in sometimes millions of data points, algorithmically infer those needed "zero catch" records.
+
+## General public needs
+Though the `afscgap` Python package makes GAP catch data more accessible to developer-lead work, the size and complexity of this dataset requires non-trivial engineering for comparative analysis between species, years, and/or geographic areas [@notebook]. Therefore, this project also offers visualization tools sitting on top of `afscgap` to begin investigations. Employing information and game design, this tool lowers usability barriers to address audiences of broad technical sophistication. Furthermore, it offers both Python code generation as a bridge to `afscgap` and CSV export for those at home in other tools.
+
+# Methods
+
+This project's design aims to improve accessibility of NOAA AFSC GAP catch data, democratizing developer access to the sophisticated methods required to interact with these data and offering inclusive approachable tools to kickstart analysis.
+
+## API design
+
[…]
