### 2023-03-24 6703988
Reorg structure to more closely align with JOSS guidelines.
