+++ inst/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-The National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center (NOAA AFSC) runs the [Groundfish Assessment Program](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program) which produces marine longitudinal catch data [@afscgap]. These "hauls" report where species are found and in what quantities, empowering biodiversity research and fisheries management [@pacificcod]. Increasing accessibility of these important data through a suite of tools for individuals of diverse programming experience, Pyafscgap.org offers not just easier access to the REST API through ORDS compilation but provides both memory-efficient algorithms for "zero-catch inference" and interactive visual analytics tools [@ords]. Altogether, this toolset supports investigatory tasks not easily executable using the API service alone and, leveraging game and information design, offers these data to a broader audience.
+The National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center (NOAA AFSC RACE) runs the [Groundfish Assessment Program](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program) which produces marine longitudinal catch data via bottom trawl surveys [@afscgap]. These "hauls" report where species are found and in what quantities, data empowering biodiversity research and fisheries management [@example]. Increasing accessibility of these important data (called RACEBASE) through a suite of tools for individuals of diverse programming experience, Pyafscgap.org offers not just easier access to the REST API through ORDS compilation but provides both memory-efficient algorithms for "zero-catch inference" and interactive visual analytics tools [@inport; @ords]. Altogether, this toolset supports investigatory tasks not easily executable using the API service alone and, leveraging game and information design, offers these data to a broader audience.
 
 # Statement of need
 Pyafscgap.org reduces barriers for use of AFSC GAP data, offering open source solutions for addressing the dataset's presence-only nature, use of proprietary databases, size, and complexity [@inport].
@@ -41,7 +41,7 @@ Though the `afscgap` Python package makes GAP catch data more accessible to deve
 This project's design aims to improve accessibility of NOAA AFSC GAP catch data, democratizing developer access to the sophisticated methods required to interact with these data and offering inclusive approachable tools to kickstart analysis.
 
 ## Lazy querying
-Starting with the `afscgap` library, [lazy iterators]() increase accessibility of the data by encapsulating logic for memory-efficient pagination and "data munging" behind a familiar interface [@lazy]. Furthermore, to support zero catch data, [decorators]() adapt diverse structures to common interfaces to free client code from understanding the full complexities of `afscgap`'s type system [@decorators]. Finally, offering a single function entry-point into the library with keywords for complex use, this "facade" approach allows the user to interact with these systems without requiring client code to reflect deep understanding of the library's mechanics, a goal furthered by compilation of "standard" Python types to Oracle REST Data Service queries [@facade].
+Starting with the `afscgap` library, lazy "generator iterables" increase accessibility of the data by encapsulating logic for memory-efficient pagination and "data munging" behind a familiar interface [@lazy]. Furthermore, to support zero catch data, decorators adapt diverse structures to common interfaces to free client code from understanding the full complexities of `afscgap`'s type system [@decorators]. Finally, offering a single function entry-point into the library with keywords for complex use, this "facade" approach allows the user to interact with these systems without requiring client code to reflect deep understanding of the library's mechanics, a goal furthered by compilation of "standard" Python types to Oracle REST Data Service queries [@facade].
 
 ## Zero catch inference
 "Negative" or "zero catch" uses the following proceedure:
@@ -65,7 +65,7 @@ However, building competency in this sophisticated interface presents user exper
[…]
