+++ inst/paper.bib
@@ -1,0 +1,109 @@
+@misc{afscgap,
+  title={Groundfish Assessment Program},
+  url={https://www.fisheries.noaa.gov/contact/groundfish-assessment-program},
+  author={NOAA Fisheries},
+  publisher={National Oceanic and Atmospheric Administration}
+}
+
+@misc{decorator,
+  title={Decorator},
+  url={https://refactoring.guru/design-patterns/decorator},
+  journal={Refactoring.Guru},
+  publisher={Refactoring.Guru},
+  author={Shvets, Alexander},
+  year={2023}
+}
+
+@article{example,
+  title={Coral in Alaska: distribution, abundance, and species associations}, volume={471}, DOI={10.1023/a:1016528631593},
+  number={1/3},
+  journal={Hydrobiologia},
+  author={Heifetz, Jonathan},
+  year={2002},
+  pages={19–28}
+}
+
+@misc{facade,
+  title={Facade},
+  url={https://refactoring.guru/design-patterns/facade},
+  journal={Refactoring.Guru},
+  publisher={Refactoring.Guru},
+  author={Shvets, Alexander},
+  year={2023}
+}
+
+@misc{flatfile,
+  title={Python NOAA AFSC GAP Tools (Downloads)},
+  url={https://pyafscgap.org/#downloads},
+  journal={AFSC GAP Tools for Python},
+  publisher={University of California Berkeley},
+  author={Pottinger, A S},
+  year={2023},
+  month={Mar}
+}
+
+@misc{geohash,
+  title={Geohash.org is public!},
+  url={https://blog.labix.org/2008/02/26/geohashorg-is-public},
+  journal={Labix Blog},
+  publisher={Labix},
+  author={Niemeyer, Gustavo},
+  year={2008},
+  month={Feb}
+}
+
+@misc{inport,
+  title={AFSC/race/gap: Racebase database},
+  url={https://www.fisheries.noaa.gov/inport/item/22008.},
+  journal={InPort},
+  publisher={Fisheries Information System},
+  author={Kenney, Heather and Roberson, Nancy},
+  year={2022},
+  month={Aug}
+}
+
+@misc{lazy,
+  title={Lazy looping in python: Making and using generators and iterators},
+  url={https://pycon2019.trey.io/index.html},
+  journal={Pycon 2019},
+  publisher={Python Software Foundation},
+  author={Hunner, Trey},
+  year={2019}
+} 
+
+@misc{notebook,
+  title={Cod AFSC GAP Example},
+  url={https://hub.gke2.mybinder.org/user/schmidtdse-afscgap-ia1m4wd3/notebooks/index.ipynb}, 
+  journal={MyBinder},
+  publisher={The Binder Team},
+  author={Pottinger, A S},
+  year={2023},
+  month={Mar}
+} 
+
+@misc{ords,
+  title={Oracle Rest Data Services},
+  url={https://docs.oracle.com/en/database/oracle/oracle-rest-data-services/},
+  journal={Oracle Help Center},
+  publisher={Oracle},
+  year={2022},
+  month={Dec}
+}
+
+@misc{readme,
+  title={AFSC GAP Viz README.md},
+  url={https://github.com/SchmidtDSE/afscgap/blob/main/afscgapviz/README.md},
+  journal={GitHub},
+  publisher={GitHub, Inc.},
+  author={Pottinger, A S},
+  year={2023},
+  month={Mar}
+}
+
+@misc{thinkaloud,
+  title={Using the "Thinking-aloud" method in cognitive interface design},
+  author={Lewis, Clayton},
+  year={1982},
+  publisher={IBM TJ Watson Research Center Yorktown Heights, NY},
+  url={https://dominoweb.draco.res.ibm.com/2513e349e05372cc852574ec0051eea4.html}
+}
