+++ inst/paper.bib
@@ -5,6 +5,16 @@
   publisher={National Oceanic and Atmospheric Administration}
 }
 
+@misc{brown,
+  title={Super Mario 3D World's 4 Step Level Design},
+  url={https://www.youtube.com/watch?v=dBmIkEvEBtA},
+  journal={Game Maker's Toolkit},
+  publisher={YouTube},
+  author={Brown, Mark},
+  year={2015},
+  month={Mar}
+}
+
 @misc{decorator,
   title={Decorator},
   url={https://refactoring.guru/design-patterns/decorator},
