+++ inst/paper.bib
@@ -114,6 +114,6 @@
   title={Using the "Thinking-aloud" method in cognitive interface design},
   author={Lewis, Clayton},
   year={1982},
-  publisher={IBM TJ Watson Research Center Yorktown Heights, NY},
+  publisher={IBM TJ Watson Research Center},
   url={https://dominoweb.draco.res.ibm.com/2513e349e05372cc852574ec0051eea4.html}
 }
