+++ inst/paper.bib
@@ -62,6 +62,15 @@
   month={Feb}
 }
 
+@misc{hayashida,
+  title={Thinking in 3D: The development of Super Mario 3D land},
+  url={https://www.gdcvault.com/play/1015833/Thinking-In-3D-The-Development},
+  journal={GDC Vault},
+  publisher={Informa PLC},
+  author={Hayashida, Koichi},
+  year={2012}
+}
+
 @misc{inport,
   title={AFSC/race/gap: Racebase database},
   url={https://www.fisheries.noaa.gov/inport/item/22008.},
