+++ inst/paper.bib
@@ -62,13 +62,14 @@
   month={Feb}
 }
 
-@misc{hayashida,
-  title={Thinking in 3D: The development of Super Mario 3D land},
-  url={https://www.gdcvault.com/play/1015833/Thinking-In-3D-The-Development},
-  journal={GDC Vault},
-  publisher={Informa PLC},
-  author={Hayashida, Koichi},
-  year={2012}
+ @misc{hayashida,
+  title={The Structure of Fun},
+  url={https://www.gamedeveloper.com/design/the-structure-of-fun-learning-from-i-super-mario-3d-land-i-s-director},
+  journal={Game Developer},
+  publisher={Informa},
+  author={Nutt, Christian and Hayashida, Koichi},
+  year={2012},
+  month={Apr}
 }
 
 @misc{inport,
