### 2023-03-24 e437378
Add diagrams and select library citations.
