+++ inst/paper.bib
@@ -15,12 +15,39 @@
   month={Mar}
 }
 
+@misc{colorbrewer,
+  title={Colorbrewer 2.0},
+  url={https://colorbrewer2.org},
+  journal={ColorBrewer},
+  publisher={The Pennsylvania State University},
+  author={Brewer, Cynthia and Harrower, Mark and Sheesley, Ben and Woodruff, Andy and Heyman, David},
+  year={2013}
+} 
+
 @misc{decorator,
   title={Decorator},
   url={https://refactoring.guru/design-patterns/decorator},
   journal={Refactoring.Guru},
   publisher={Refactoring.Guru},
   author={Shvets, Alexander},
+  year={2023}
+}
+
+@misc{diagrams,
+  title={Jgraph/Drawio: Draw.io is a JavaScript, client-side editor for general diagramming and whiteboarding},
+  url={https://github.com/jgraph/drawio},
+  journal={GitHub},
+  publisher={GitHub, Inc}, 
+  uthor={JGraph Ltd and draw.io AG},
+  year={2023}
+}
+
+@misc{d3,
+  title={Data-Driven Documents 7.8.2},
+  url={https://d3js.org/},
+  journal={D3.js},
+  publisher={Mike Bostock},
+  author={Bostock, Mike},
   year={2023}
 }
 
@@ -110,6 +137,15 @@
   month={Dec}
 }
 
+@misc{papa,
+  title={Papa Parse - Powerful CSV Parser for JavaScript},
+  url={https://www.papaparse.com/},
+  journal={Papa Parse},
+  publisher={Matt Holt},
+  author={Holt, Matt},
+  year={2023}
+}
+
 @misc{readme,
   title={AFSC GAP Viz README.md},
   url={https://github.com/SchmidtDSE/afscgap/blob/main/afscgapviz/README.md},
@@ -118,6 +154,15 @@
   author={Pottinger, A S},
   year={2023},
   month={Mar}
+}
+
+@misc{requests,
+  title={HTTP for humans},
+  url={https://docs.python-requests.org/en/latest/index.html},
+  journal={Requests},
+  publisher={Requests Project},
+  author={Reitz, Kenneth},
+  year={2023}
 }
 
 @misc{thinkaloud,
