### 2023-03-24 ed81a2c
Expand citations a bit.
