+++ inst/paper.bib
@@ -47,7 +47,7 @@
   url={https://d3js.org/},
   journal={D3.js},
   publisher={Mike Bostock},
-  author={Bostock, Mike},
+  author={Bostock, Mike and D3 Contributors},
   year={2023}
 }
 
@@ -69,6 +69,15 @@
   year={2023}
 }
 
+@misc{flask,
+  title={Welcome to Flask},
+  url={https://flask.palletsprojects.com/en/2.2.x/},
+  journal={Pallets Projects},
+  publisher={Pallets},
+  author={Mönnich, Adrian and Ronacher, Armin and Lord, David and Li, Grey and Bronson, Joshua and Unterwaditzer, Markus and Jones, Philip and Flask Contributors},
+  year={2023}
+}
+
 @misc{flatfile,
   title={Python NOAA AFSC GAP Tools (Downloads)},
   url={https://pyafscgap.org/#downloads},
@@ -87,6 +96,15 @@
   author={Niemeyer, Gustavo},
   year={2008},
   month={Feb}
+}
+
+@misc{geolib,
+  title={Joyanujoy/geolib: Python geohash library},
+  url={https://github.com/joyanujoy/geolib},
+  journal={GitHub},
+  publisher={GitHub Inc},
+  author={Joy, Anu and Rivard, Éloi},
+  year={2021}
 }
 
  @misc{hayashida,
@@ -171,4 +189,13 @@
   year={1982},
   publisher={IBM TJ Watson Research Center},
   url={https://dominoweb.draco.res.ibm.com/2513e349e05372cc852574ec0051eea4.html}
+}
+
+@misc{toolz,
+  title={Pytoolz API Documentation},
+  url={https://toolz.readthedocs.io/en/latest/},
+  journal={Read the Docs},
+  publisher={Read the Docs, Inc},
+  author={Rocklin, Matthew and Jacobsen, John and Toolz Contributors},
+  year={2022}
 }
