### 2023-03-24 ad4045d
Add missing docstrings for query_util.
