+++ afscgapviz/build_database.py
@@ -12,7 +12,7 @@ See build_database.sh for example.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import contextlib
 import re
