+++ afscgapviz/data_util.py
@@ -7,7 +7,7 @@ data representations.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import typing
 
