+++ afscgap/inference.py
@@ -5,7 +5,7 @@ Tools for inferring missing, negative, or zero catch records.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import copy
 import csv
