+++ afscgap/client.py
@@ -6,7 +6,7 @@ with the NOAA-run AFSC GAP API.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import copy
 import json
