+++ afscgapviz/test/test_sql_util.py
@@ -5,7 +5,7 @@ Tests for afscgap backend SQL utility functions.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import unittest
 
