+++ afscgapviz/survey_util.py
@@ -4,7 +4,7 @@
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import sqlite3
 import typing
