+++ afscgapviz/test/test_data_util.py
@@ -5,7 +5,7 @@ Tests for data support functions.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import unittest
 
