+++ afscgapviz/test/test_afscgapviz.py
@@ -5,7 +5,7 @@ Tests for the afscgap web based visualization.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import unittest
 
