+++ afscgap/test/test_tools.py
@@ -5,7 +5,7 @@ Convienence functions for testing afscgap.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 import json
 import os
