+++ afscgap/cursor.py
@@ -5,7 +5,7 @@ Interfaces for cursor objects which iterate over real or inferred records.
 for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
-LICENSE.txt.
+LICENSE.md.
 """
 
 import queue
