+++ inst/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-The National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center (NOAA AFSC RACE) runs the [Groundfish Assessment Program](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program) which produces marine longitudinal catch data via bottom trawl surveys [@afscgap]. These "hauls" report where species are found and in what quantities, data empowering biodiversity research and fisheries management [@example]. Increasing accessibility of these important data (called RACEBASE) through a suite of tools for individuals of diverse programming experience, Pyafscgap.org offers not just easier access to the REST API through ORDS compilation but provides both memory-efficient algorithms for "zero-catch inference" and interactive visual analytics tools [@inport; @ords]. Altogether, this toolset supports investigatory tasks not easily executable using the API service alone and, leveraging game and information design, offers these data to a broader audience.
+The National Oceanic and Atmospheric Administration's Alaska Fisheries Science Center (NOAA AFSC RACE) runs the [Groundfish Assessment Program](https://www.fisheries.noaa.gov/contact/groundfish-assessment-program) which produces longitudinal catch data via bottom trawl surveys [@afscgap]. These "hauls" report where species are found and in what quantities, data empowering biodiversity research and fisheries management [@example]. Increasing accessibility of these important data (called RACEBASE) through a suite of tools for individuals of diverse programming experience, Pyafscgap.org offers not just easier access to the REST API through ORDS compilation but provides both memory-efficient algorithms for "zero-catch inference" and interactive visual analytics tools [@inport; @ords]. Altogether, this toolset supports investigatory tasks not easily executable using the API service alone and, leveraging game and information design, offers these data to a broader audience.
 
 # Statement of need
 Pyafscgap.org reduces barriers for use of AFSC GAP data, offering open source solutions for addressing the dataset's presence-only nature, use of proprietary databases, size, and complexity [@inport].
