### 2023-03-25 fcd1342
Respond to @gizarp feedback for #41.
