### 2023-03-25 512ddbe
Mention GAP acronym.
